consumer/requirements.txt
pika==1.1.0
six==1.13.0
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
numpy==1.19.1
pymongo==3.11.2
Flask_Cors==3.0.4
dlib==19.17.0
requests==2.21.0
Flask==1.1.2
scipy==1.5.2
matplotlib==3.1.2
Pillow==8.1.0
